move enums out of components and .tsx -> .ts
