--- apps/frontend/.npmrc
@@ -1,1 +1,0 @@
-legacy-peer-deps=true
