+++ apps/frontend/enums/Category.ts
