--- apps/frontend/src/components/User.tsx
@@ -1,6 +1,0 @@
-export type User = {
-  id: number
-  username: string
-  email: string
-  password: string
-}
