+++ apps/frontend/src/pages/signup.tsx
@@ -1,5 +1,5 @@
 import UserForm from "../components/forms/UserForm";
-import { UserManagement } from "../components/enums/UserManagement";
+import { UserManagement } from "../../enums/UserManagement";
 import { signIn } from "next-auth/react";
 
 export default function Signup() {
